+++ test/bones_api_enitity_db_postgres_test.dart
@@ -31,28 +31,34 @@ class PostgresTestConfig extends APITestConfigDockerPostgreSQL {
 }
 
 Future<void> main() async {
-  await _runTest(true, false, false);
-  await _runTest(false, false, false);
+  await _runTest(true, false, false, false);
+  await _runTest(false, false, false, false);
 
-  await _runTest(true, true, false);
-  await _runTest(true, true, true);
+  await _runTest(true, true, false, false);
+  await _runTest(true, true, true, false);
+  await _runTest(true, true, true, true);
 }
 
-Future<bool> _runTest(
-        bool useReflection, bool generateTables, bool checkTables) =>
+Future<bool> _runTest(bool useReflection, bool generateTables, bool checkTables,
+        bool populateSource) =>
     runAdapterTests(
       'PostgreSQL',
       PostgresTestConfig(
           generateTables: generateTables, checkTables: checkTables),
-      (provider, dbPort, dbConfig) => DBPostgreSQLAdapter(
-        dbName,
-        dbUser,
-        password: dbPass,
-        port: dbPort,
-        parentRepositoryProvider: provider,
-        generateTables: generateTables,
-        checkTables: checkTables,
-      ),
+      (provider, dbPort, dbConfig) {
+        var populate = dbConfig?['populate'] as Map?;
+        return DBPostgreSQLAdapter(
+          dbName,
+          dbUser,
+          password: dbPass,
+          port: dbPort,
+          parentRepositoryProvider: provider,
+          generateTables: generateTables,
+          checkTables: checkTables,
+          populateSource: populate?['source'],
+          populateSourceVariables: populate?['variables'],
+        );
+      },
       (provider, dbPort, dbConfig) =>
           DBObjectMemoryAdapter(parentRepositoryProvider: provider),
       '"',
@@ -60,4 +66,5 @@ Future<bool> _runTest(
       entityByReflection: useReflection,
       generateTables: generateTables,
       checkTables: checkTables,
+      populateSource: populateSource,
     );