+++ test/bones_api_entity_db_mysql_test.dart
@@ -33,28 +33,34 @@ class MySQLTestConfig extends APITestConfigDockerMySQL {
 }
 
 Future<void> main() async {
-  await _runTest(false, false, false);
-  await _runTest(true, false, false);
+  await _runTest(false, false, false, false);
+  await _runTest(true, false, false, false);
 
-  await _runTest(true, true, false);
-  await _runTest(true, true, true);
+  await _runTest(true, true, false, false);
+  await _runTest(true, true, true, false);
+  await _runTest(true, true, true, true);
 }
 
-Future<bool> _runTest(
-        bool useReflection, bool generateTables, bool checkTables) =>
+Future<bool> _runTest(bool useReflection, bool generateTables, bool checkTables,
+        bool populateSource) =>
     runAdapterTests(
       'MySQL',
       MySQLTestConfig(generateTables: generateTables, checkTables: checkTables),
-      (provider, dbPort, dbConfig) => DBMySQLAdapter(
-        dbName,
-        dbUser,
-        password: dbPass,
-        host: '127.0.0.1',
-        port: dbPort,
-        parentRepositoryProvider: provider,
-        generateTables: generateTables,
-        checkTables: checkTables,
-      ),
+      (provider, dbPort, dbConfig) {
+        var populate = dbConfig?['populate'] as Map?;
+        return DBMySQLAdapter(
+          dbName,
+          dbUser,
+          password: dbPass,
+          host: '127.0.0.1',
+          port: dbPort,
+          parentRepositoryProvider: provider,
+          generateTables: generateTables,
+          checkTables: checkTables,
+          populateSource: populate?['source'],
+          populateSourceVariables: populate?['variables'],
+        );
+      },
       (provider, dbPort, dbConfig) =>
           DBObjectMemoryAdapter(parentRepositoryProvider: provider),
       '`',
@@ -62,4 +68,5 @@ Future<bool> _runTest(
       entityByReflection: useReflection,
       generateTables: generateTables,
       checkTables: checkTables,
+      populateSource: populateSource,
     );